@@ -1,54 +1,44 @@
 {
-	"name": "debt-destroyer",
-	"version": "1.0.0",
-	"private": true,
-	"dependencies": {
-		"@material-ui/core": "^4.2.1",
-		"@material-ui/icons": "^4.2.1",
-		"@types/jest": "24.0.15",
-		"@types/nanoid": "^2.0.0",
-		"@types/node": "12.6.8",
-		"@types/react": "16.8.23",
-		"@types/react-dom": "16.8.5",
-		"date-fns": "^1.30.1",
-		"lodash": "^4.17.15",
-		"nanoid": "^2.0.3",
-		"query-string": "^6.8.2",
-		"react": "^16.8.6",
-		"react-dom": "^16.8.6",
-		"react-scripts": "3.0.1",
-		"recharts": "^1.6.2",
-		"typescript": "3.5.3"
-	},
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject"
-	},
-	"eslintConfig": {
-		"extends": "react-app"
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	},
-	"devDependencies": {
-		"@types/lodash": "^4.14.136",
-		"@types/recharts": "^1.1.20",
-		"@types/styled-components": "^4.1.18",
-		"aws-sdk": "^2.507.0",
-		"dotenv": "^8.0.0",
-		"glob": "^7.1.4",
-		"mime-types": "^2.1.24",
-		"prettier": "^1.18.2"
-	}
+  "name": "debt-destroyer",
+  "version": "2.0.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc -b && vite build",
+    "preview": "vite preview",
+    "test": "vitest"
+  },
+  "dependencies": {
+    "@radix-ui/react-dialog": "^1.1.15",
+    "@radix-ui/react-label": "^2.1.8",
+    "@radix-ui/react-radio-group": "^1.3.8",
+    "@radix-ui/react-slot": "^1.2.4",
+    "@radix-ui/react-tabs": "^1.1.13",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "date-fns": "^3.6.0",
+    "lodash": "^4.17.21",
+    "lucide-react": "^0.563.0",
+    "nanoid": "^5.0.9",
+    "query-string": "^9.1.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "recharts": "^2.15.0",
+    "tailwind-merge": "^3.4.0"
+  },
+  "devDependencies": {
+    "@tailwindcss/postcss": "^4.1.18",
+    "@types/lodash": "^4.17.13",
+    "@types/node": "^22.10.7",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
+    "@vitejs/plugin-react": "^4.3.4",
+    "autoprefixer": "^10.4.24",
+    "postcss": "^8.5.6",
+    "tailwindcss": "^4.1.18",
+    "typescript": "^5.7.3",
+    "vite": "^6.0.7",
+    "vitest": "^2.1.8"
+  }
 }
