@@ -1,139 +1,98 @@
-import * as React from 'react';
+import { useMemo } from 'react'
+import { format, addMonths } from 'date-fns'
 import {
-	calculateDebts,
-	DEBT_PAYOFF_METHODS,
-	IRepaymentSchedule
-} from '../utils';
-import formatDate from 'date-fns/format';
-import addMonths from 'date-fns/add_months';
-import Table from '@material-ui/core/Table';
-import TableRow from '@material-ui/core/TableRow';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableHead from '@material-ui/core/TableHead';
-import { IDebt } from './app/app';
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from '@/components/ui/table'
+import {
+  calculateDebts,
+  DEBT_PAYOFF_METHODS,
+  IRepaymentSchedule,
+  IDebt
+} from '../utils'
 
-interface IProps {
-	debtData: IRepaymentSchedule;
-	debts: IDebt[];
-	extraContributions: string;
-	debtPayoffMethod: DEBT_PAYOFF_METHODS;
-}
-
-interface IState {
-	debtData: IRepaymentSchedule;
-	fiftyExtraScenario: IRepaymentSchedule;
-	oneHundredFiftyExtraScenario: IRepaymentSchedule;
+interface InsightsProps {
+  debtData: IRepaymentSchedule
+  debts: IDebt[]
+  extraContributions: string
+  debtPayoffMethod: DEBT_PAYOFF_METHODS
 }
 
 function getTotalInterestPaid(debtData: IRepaymentSchedule): number {
-	return debtData.months.reduce((acc, month) => {
-		return (
-			acc +
-			Object.values(month.values).reduce((acc, value) => {
-				if (!value.interestPaid) {
-					return acc;
-				}
-
-				return acc + value.interestPaid;
-			}, 0)
-		);
-	}, 0);
+  return debtData.months.reduce((acc, month) => {
+    return (
+      acc +
+      Object.values(month.values).reduce((acc, value) => {
+        if (!value.interestPaid) {
+          return acc
+        }
+        return acc + value.interestPaid
+      }, 0)
+    )
+  }, 0)
 }
 
 function getDebtPayoffDate(debtData: IRepaymentSchedule): string {
-	return formatDate(
-		addMonths(
-			new Date(),
-			debtData.months[debtData.months.length - 1].month
-		),
-		'MMM, YYYY'
-	);
+  return format(
+    addMonths(new Date(), debtData.months[debtData.months.length - 1].month),
+    'MMM, yyyy'
+  )
 }
 
-export default class Insights extends React.Component<IProps, IState> {
-	state = {
-		debtData: this.props.debtData,
-		fiftyExtraScenario: calculateDebts({
-			debtMethod: this.props.debtPayoffMethod,
-			extraContributions:
-				parseInt(this.props.extraContributions, 10) + 50,
-			debts: this.props.debts
-		}),
-		oneHundredFiftyExtraScenario: calculateDebts({
-			debtMethod: this.props.debtPayoffMethod,
-			extraContributions:
-				parseInt(this.props.extraContributions, 10) + 150,
-			debts: this.props.debts
-		})
-	};
+export default function Insights({
+  debtData,
+  debts,
+  extraContributions,
+  debtPayoffMethod
+}: InsightsProps) {
+  const fiftyExtraScenario = useMemo(
+    () =>
+      calculateDebts({
+        debtMethod: debtPayoffMethod,
+        extraContributions: parseInt(extraContributions, 10) + 50,
+        debts
+      }),
+    [debtPayoffMethod, extraContributions, debts]
+  )
 
-	static getDerivedStateFromProps(props, state) {
-		if (props.debtData.guid === state.debtData.guid) {
-			return null;
-		}
+  const oneHundredFiftyExtraScenario = useMemo(
+    () =>
+      calculateDebts({
+        debtMethod: debtPayoffMethod,
+        extraContributions: parseInt(extraContributions, 10) + 150,
+        debts
+      }),
+    [debtPayoffMethod, extraContributions, debts]
+  )
 
-		return {
-			debtData: props.debtData,
-			fiftyExtraScenario: calculateDebts({
-				debtMethod: props.debtPayoffMethod,
-				extraContributions: parseInt(props.extraContributions, 10) + 50,
-				debts: props.debts
-			}),
-			oneHundredFiftyExtraScenario: calculateDebts({
-				debtMethod: props.debtPayoffMethod,
-				extraContributions:
-					parseInt(props.extraContributions, 10) + 150,
-				debts: props.debts
-			})
-		};
-	}
-
-	render() {
-		const { debtData } = this.props;
-		return (
-			<Table>
-				<TableHead>
-					<TableRow>
-						<TableCell />
-						<TableCell>Current scenario</TableCell>
-						<TableCell>Extra $50 a month</TableCell>
-						<TableCell>Extra $150 a month</TableCell>
-					</TableRow>
-				</TableHead>
-				<TableBody>
-					<TableRow>
-						<TableCell>Pay off date</TableCell>
-						<TableCell>{getDebtPayoffDate(debtData)}</TableCell>
-						<TableCell>
-							{getDebtPayoffDate(this.state.fiftyExtraScenario)}
-						</TableCell>
-						<TableCell>
-							{getDebtPayoffDate(
-								this.state.oneHundredFiftyExtraScenario
-							)}
-						</TableCell>
-					</TableRow>
-					<TableRow>
-						<TableCell>Total interest paid</TableCell>
-						<TableCell>
-							${getTotalInterestPaid(debtData).toFixed(2)}
-						</TableCell>
-						<TableCell>
-							$
-							{getTotalInterestPaid(
-								this.state.fiftyExtraScenario
-							).toFixed(2)}
-						</TableCell>
-						<TableCell>
-							$
-							{getTotalInterestPaid(
-								this.state.oneHundredFiftyExtraScenario
-							).toFixed(2)}
-						</TableCell>
-					</TableRow>
-				</TableBody>
-			</Table>
-		);
-	}
+  return (
+    <Table>
+      <TableHeader>
+        <TableRow>
+          <TableHead />
+          <TableHead>Current scenario</TableHead>
+          <TableHead>Extra $50 a month</TableHead>
+          <TableHead>Extra $150 a month</TableHead>
+        </TableRow>
+      </TableHeader>
+      <TableBody>
+        <TableRow>
+          <TableCell className="font-medium">Pay off date</TableCell>
+          <TableCell>{getDebtPayoffDate(debtData)}</TableCell>
+          <TableCell>{getDebtPayoffDate(fiftyExtraScenario)}</TableCell>
+          <TableCell>{getDebtPayoffDate(oneHundredFiftyExtraScenario)}</TableCell>
+        </TableRow>
+        <TableRow>
+          <TableCell className="font-medium">Total interest paid</TableCell>
+          <TableCell>${getTotalInterestPaid(debtData).toFixed(2)}</TableCell>
+          <TableCell>${getTotalInterestPaid(fiftyExtraScenario).toFixed(2)}</TableCell>
+          <TableCell>${getTotalInterestPaid(oneHundredFiftyExtraScenario).toFixed(2)}</TableCell>
+        </TableRow>
+      </TableBody>
+    </Table>
+  )
 }
