@@ -1,4 +1,8 @@
+export const MONTHS_PER_YEAR = 12
+export const MINIMUM_PRINCIPAL_RATE = 0.01
+export const DEBOUNCE_MS = 300
+
 export const errorTemplate = {
-	error: false,
-	message: ''
-};
+  error: false,
+  message: ''
+}
